third_party/blink/web_tests/platform/generic/external/wpt/css/css-align/parsing/place-self-shorthand-expected.txt

This is a testharness.js-based test.
PASS e.style['place-self'] = "normal" should set align-self
PASS e.style['place-self'] = "normal" should set justify-self
PASS e.style['place-self'] = "normal" should not set unrelated longhands
PASS e.style['place-self'] = "first baseline" should set align-self
PASS e.style['place-self'] = "first baseline" should set justify-self
PASS e.style['place-self'] = "first baseline" should not set unrelated longhands
FAIL e.style['place-self'] = "last baseline flex-start" should set align-self assert_equals: align-self should be canonical expected "last baseline" but got ""
FAIL e.style['place-self'] = "last baseline flex-start" should set justify-self assert_equals: justify-self should be canonical expected "flex-start" but got ""
FAIL e.style['place-self'] = "last baseline flex-start" should not set unrelated longhands assert_true: expected true got false
PASS e.style['place-self'] = "unsafe self-start stretch" should set align-self
PASS e.style['place-self'] = "unsafe self-start stretch" should set justify-self
PASS e.style['place-self'] = "unsafe self-start stretch" should not set unrelated longhands
Harness: the test ran to completion.
